Feat: created custom message for "entity not found" case

+++ ProductManagement/ProductManagement.Application/UseCases/Commands/DeleteProduct/DeleteProductCommandHandler.cs
@@ -12,7 +12,9 @@ public class DeleteProductCommandHandler(
         var products = await productRepository
             .FindByCondition(product => product.Id == request.ProductId,
             false, cancellationToken);
-        var product = products.First();
+        var product = products.FirstOrDefault();
+        if (product is null)
+            throw new InvalidOperationException($"Product with id {request.ProductId} not found");
         
         await productRepository.Delete(product, cancellationToken);
